Move evict.h back to include/
